+++ src/components/KymaModules/deleteModulesHelpers.tsx
@@ -5,6 +5,7 @@ import {
   ModuleTemplateListType,
 } from './support';
 import { PostFn } from 'shared/hooks/BackendAPI/usePost';
+import postForCommunityResources from 'components/KymaModules/components/CommunityModulesHelpers';
 
 interface Counts {
   [key: string]: number;
@@ -276,28 +277,6 @@ export const deleteCrResources = async (
     return 'Error while deleting Custom Resource';
   }
 };
-
-export default async function postForCommunityResources(
-  post: PostFn,
-  link: string,
-) {
-  if (!link) {
-    console.error('No link provided for community resource');
-    return false;
-  }
-
-  try {
-    const response = await post('/modules/community-resource', { link });
-    if (response?.length) {
-      return response;
-    }
-    console.error('Empty or invalid response:', response);
-    return false;
-  } catch (error) {
-    console.error('Error fetching data:', error);
-    return false;
-  }
-}
 
 export const getCommunityResourceUrls = (resources: any) => {
   if (!resources?.length) return [];
