+++ src/components/KymaModules/support.ts
@@ -117,13 +117,6 @@ export type ModuleReleaseMetaListType = {
   items: ModuleReleaseMetas[];
 };
 
-export const getModuleName = (moduleTemplate: ModuleTemplateType): string => {
-  return (
-    moduleTemplate.metadata?.labels['operator.kyma-project.io/module-name'] ??
-    moduleTemplate.spec.moduleName
-  );
-};
-
 export const getResourcePath = (resource: any) => {
   if (!resource) return '';
 
@@ -199,6 +192,13 @@ export const findModuleTemplate = (
   );
 
   return moduleWithInfo ?? moduleTemplateWithoutInfo;
+};
+
+export const getModuleName = (moduleTemplate: ModuleTemplateType): string => {
+  return (
+    moduleTemplate.metadata?.labels['operator.kyma-project.io/module-name'] ??
+    moduleTemplate.spec.moduleName
+  );
 };
 
 export const setChannel = (
